api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
qdrant-client==1.11.3
python-dotenv==1.0.1
numpy==2.1.2
tiktoken==0.7.0
docx2txt==0.8
